Create README.md
This README file introduces the Image-to-Text-and-Translate project and explains how to install, run, and use it. This project is an application that can convert images into text and translate the English texts into Persian if needed. This commit is intended to provide a helpful and detailed documentation for the project and its users.